* moving loading indicators to their own subdir
