+++ lib/components/AccountGovernanceClaims.jsx
@@ -15,9 +15,9 @@ import { CONTRACT_ADDRESSES, DEFAULT_TOKEN_PRECISION, SECONDS_PER_DAY } from 'li
 import { isSelfAtom } from 'lib/components/AccountUI'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
 import { AddTokenToMetaMaskButton } from 'lib/components/AddTokenToMetaMaskButton'
-import { IndexUILoader } from 'lib/components/IndexUILoader'
+import { IndexUILoader } from 'lib/components/loaders/IndexUILoader'
 import { PoolCurrencyIcon } from 'lib/components/PoolCurrencyIcon'
-import { ThemedClipLoader } from 'lib/components/ThemedClipLoader'
+import { ThemedClipLoader } from 'lib/components/loaders/ThemedClipLoader'
 import { useAccountQuery } from 'lib/hooks/useAccountQuery'
 import { usePools } from 'lib/hooks/usePools'
 import { useSendTransaction } from 'lib/hooks/useSendTransaction'
