+++ lib/components/AccountRewards.jsx
@@ -12,7 +12,7 @@ import { EtherscanTxLink } from 'lib/components/EtherscanTxLink'
 import { PoolCurrencyIcon } from 'lib/components/PoolCurrencyIcon'
 import { PoolNumber } from 'lib/components/PoolNumber'
 import { PoolCountUp } from 'lib/components/PoolCountUp'
-import { ThemedClipLoader } from 'lib/components/ThemedClipLoader'
+import { ThemedClipLoader } from 'lib/components/loaders/ThemedClipLoader'
 import { usePlayerDrips } from 'lib/hooks/usePlayerDrips'
 import { usePool } from 'lib/hooks/usePool'
 import { usePools } from 'lib/hooks/usePools'
