+++ lib/components/IndexUI.jsx
@@ -7,9 +7,7 @@ import { ANIM_BANNER_VARIANTS } from 'lib/constants/framerAnimations'
 import { POOL_LIST_TABS } from 'lib/constants'
 import { useTranslation } from 'lib/../i18n'
 import { useReducedMotion } from 'lib/hooks/useReducedMotion'
-import { usePools } from 'lib/hooks/usePools'
 import { CommunityDisclaimerBanner } from 'lib/components/CommunityDisclaimerBanner'
-import { IndexUILoader } from 'lib/components/IndexUILoader'
 import { PoolLists } from 'lib/components/PoolLists'
 import { RetroactivePoolClaimBanner } from 'lib/components/RetroactivePoolClaimBanner'
 import { Tagline } from 'lib/components/Tagline'
