+++ lib/components/ManageUI.jsx
@@ -14,7 +14,7 @@ import { useContractAddresses } from 'lib/hooks/useContractAddresses'
 import { usePool } from 'lib/hooks/usePool'
 import { usePools } from 'lib/hooks/usePools'
 import { PoolActionsUI } from 'lib/components/PoolActionsUI'
-import { IndexUILoader } from 'lib/components/IndexUILoader'
+import { IndexUILoader } from 'lib/components/loaders/IndexUILoader'
 import { Tagline } from 'lib/components/Tagline'
 import { displayAmountInEther } from 'lib/utils/displayAmountInEther'
 import { numberWithCommas } from 'lib/utils/numberWithCommas'
