+++ lib/components/PoolPrizeListing.jsx
@@ -5,7 +5,7 @@ import { PRIZE_PAGE_SIZE } from 'lib/constants'
 import { useTranslation } from 'lib/../i18n'
 import { BlankStateMessage } from 'lib/components/BlankStateMessage'
 import { ButtonLink } from 'lib/components/ButtonLink'
-import { TableRowUILoader } from 'lib/components/TableRowUILoader'
+import { TableRowUILoader } from 'lib/components/loaders/TableRowUILoader'
 import { PaginationUI } from 'lib/components/PaginationUI'
 import { PrizesTable } from 'lib/components/PrizesTable'
 import { usePoolPrizesQuery } from 'lib/hooks/usePoolPrizesQuery'
