+++ lib/components/PoolShow.jsx
@@ -17,7 +17,7 @@ import { Button } from 'lib/components/Button'
 import { ButtonLink } from 'lib/components/ButtonLink'
 import { CommunityPoolDisclaimerModal } from 'lib/components/CommunityPoolDisclaimerModal'
 import { LootBoxTable } from 'lib/components/LootBoxTable'
-import { PoolShowUILoader } from 'lib/components/PoolShowUILoader'
+import { PoolShowUILoader } from 'lib/components/loaders/PoolShowUILoader'
 import { UpcomingPrizeBreakdownCard } from 'lib/components/UpcomingPrizeBreakdownCard'
 import { PageTitleAndBreadcrumbs } from 'lib/components/PageTitleAndBreadcrumbs'
 import { Meta } from 'lib/components/Meta'
