+++ lib/components/TVLAndWeeklyPrizesBanner.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react'
 
 import { Banner } from 'lib/components/Banner'
-import { BannerUILoader } from 'lib/components/BannerUILoader'
+import { BannerUILoader } from 'lib/components/loaders/BannerUILoader'
 import { useTotalPoolPrizeInterestUSD } from 'lib/hooks/useTotalPoolPrizeInterestUSD'
 import { useTotalPoolPrizeValueLockedUSD } from 'lib/hooks/useTotalPoolPrizeValueLockedUSD'
 import { numberWithCommas } from 'lib/utils/numberWithCommas'
